Refactor index.py
Add the functionality to check the new episode that was preeviously from the gogoanime. This way we can add mroe site and more adapter easier. We can also make the function where we can store the oldEpisode somewhere (e.g. file) So that we won't send any old episode that were lost when we restart the code.

TODO: Combine this to run on python flask chatbot